+++ examples/cms/src/main/java/com/github/resource4j/cms/CMSApplication.java
@@ -1,0 +1,13 @@
+package com.github.resource4j.cms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CMSApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CMSApplication.class, args);
+    }
+
+}